+++ crates/config/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "monolith_config"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+figment = { version = "0.10", features = ["test", "yaml"] }
+semver = "1.0"
+serde = { version = "1.0", features = ["derive"] }
+serde_yaml = "0.8"
+validator = { version = "0.14", features = ["derive"] }
