+++ crates/config/src/lib.rs
@@ -1,0 +1,9 @@
+mod errors;
+mod validators;
+
+pub mod constants;
+pub mod workspace;
+
+// Re-exports structs for convenience
+pub use validator::ValidationErrors;
+pub use workspace::WorkspaceConfig;
