+++ crates/workspace/Cargo.toml
@@ -4,9 +4,5 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-figment = { version = "0.10", features = ["test", "yaml"] }
+monolith_config = { path = "../config" }
 thiserror = "1.0"
-semver = "1.0"
-serde = { version = "1.0", features = ["derive"] }
-serde_yaml = "0.8"
-validator = { version = "0.14", features = ["derive"] }
